moving collisions files into collisions folder
