+++ examples/boxmodelcollisions/golovin/src/main_golcolls.cpp
@@ -48,7 +48,7 @@
 #include "runcleo/initialconditions.hpp"
 #include "runcleo/runcleo.hpp"
 #include "runcleo/sdmmethods.hpp"
-#include "superdrops/coalescence.hpp"
+#include "superdrops/collisionprobs/coalescence.hpp"
 #include "superdrops/collisionprobs/golovinprob.hpp"
 #include "superdrops/microphysicalprocess.hpp"
 #include "superdrops/motion.hpp"
