+++ examples/constthermo2d/src/main_const2D.cpp
@@ -51,7 +51,7 @@
 #include "runcleo/initialconditions.hpp"
 #include "runcleo/runcleo.hpp"
 #include "runcleo/sdmmethods.hpp"
-#include "superdrops/coalescence.hpp"
+#include "superdrops/collisionprobs/coalescence.hpp"
 #include "superdrops/collisionprobs/longhydroprob.hpp"
 #include "superdrops/condensation.hpp"
 #include "superdrops/microphysicalprocess.hpp"
