+++ libs/superdrops/CMakeLists.txt
@@ -14,9 +14,6 @@ set(LIBRARY_OUTPUT_PATH  ${CMAKE_BINARY_DIR}/lib)
 
 # Add executables and create library target
 set(SOURCES
-  "coalbure_flag.cpp"
-  "coalescence.cpp"
-  "collisionkinetics.cpp"
   "condensation.cpp"
   "impliciteuler.cpp"
   "state.cpp"
