+++ libs/superdrops/collisionprobs/coalbure_flag.cpp
@@ -4,7 +4,7 @@
  *
  * ----- CLEO -----
  * File: coalbure_flag.cpp
- * Project: superdrops
+ * Project: collisionprobs
  * Created Date: Wednesday 24th January 2024
  * Author: Clara Bayley (CB)
  * Additional Contributors:
@@ -20,6 +20,7 @@
  * in DoCoalBuRe to decide whether breakup,
  * coalescence or rebound should occur.
  */
+
 
 #include "./coalbure_flag.hpp"
