+++ libs/superdrops/collisionprobs/coalbure_flag.hpp
@@ -4,7 +4,7 @@
  *
  * ----- CLEO -----
  * File: coalbure_flag.hpp
- * Project: superdrops
+ * Project: collisionprobs
  * Created Date: Wednesday 24th January 2024
  * Author: Clara Bayley (CB)
  * Additional Contributors:
@@ -21,8 +21,9 @@
  * coalescence or rebound should occur.
  */
 
-#ifndef LIBS_SUPERDROPS_COALBURE_FLAG_HPP_
-#define LIBS_SUPERDROPS_COALBURE_FLAG_HPP_
+
+#ifndef LIBS_SUPERDROPS_COLLISIONPROBS_COALBURE_FLAG_HPP_
+#define LIBS_SUPERDROPS_COLLISIONPROBS_COALBURE_FLAG_HPP_
 
 #include <concepts>
 #include <functional>
@@ -106,4 +107,4 @@ struct TSCoalBuReFlag {
   unsigned int operator()(const double phi, const Superdrop &drop1, const Superdrop &drop2) const;
 };
 
-#endif   // LIBS_SUPERDROPS_COALBURE_FLAG_HPP_
+#endif   // LIBS_SUPERDROPS_COLLISIONPROBS_COALBURE_FLAG_HPP_
