+++ src/main.cpp
@@ -58,8 +58,8 @@
 #include "superdrops/breakup.hpp"
 #include "superdrops/breakup_nfrags.hpp"
 #include "superdrops/coalbure.hpp"
-#include "superdrops/coalbure_flag.hpp"
-#include "superdrops/coalescence.hpp"
+#include "superdrops/collisionprobs/coalescence.hpp"
+#include "superdrops/collisionprobs/coalbure_flag.hpp"
 #include "superdrops/collisionprobs/constprob.hpp"
 #include "superdrops/collisionprobs/golovinprob.hpp"
 #include "superdrops/collisionprobs/longhydroprob.hpp"
